Add files via upload
Case Study Description: 

You're consulting as a data analyst for Lariat. They've hired you to make recommendations about how to make better decisions, which they measure in terms of the costs and revenue generated by their rental car fleet.

They've provided you with raw data in four Excel worksheets in four workbooks (more on this in the next checkpoint).

They've asked you to produce the following deliverables based on the raw data and the business objectives:

An Excel model that they can use to make data-informed decisions about their fleet. This should be based on the observed costs and revenue data from their fleet in recent years.
A PowerPoint deck that provides an analysis of the problem and provides recommendations.
A live presentation in which you present your findings to a Lariat executive.